+++ server/src/nest/trips/trips.mcp.ts
@@ -7,6 +7,8 @@ import {
 import { z } from 'zod';
 import { AuthService } from '../auth/auth.service';
 import { CalendarService } from '../calendar/calendar.service';
+import { TripMembersService } from '../trip-members/trip-members.service';
+import { TripReadModelService } from '../trip-read-model/trip-read-model.service';
 import { isAddonEnabled, getCollabFeatures } from '../addons/addons.bridge';
 import { ADDON_IDS } from '../../addons';
 import { safeBroadcast, MAX_MCP_TRIP_DAYS, noAccess, hasTripPermission, permissionDenied } from '../../mcp/tools/_shared';
@@ -67,6 +69,8 @@ export class TripsMcp {
     // Appended, not inserted: the hand-wired MCP test harnesses build this
     // positionally, so an earlier slot would silently shift every one of them.
     private readonly calendar: CalendarService,
+    private readonly members: TripMembersService,
+    private readonly readModel: TripReadModelService,
   ) {}
 
   // --- TRIPS ---
@@ -204,7 +208,7 @@ export class TripsMcp {
   })
   async getTripSummary({ tripId }: { tripId: number }, ctx: McpContext) {
     if (!this.trips.canAccessTrip(tripId, ctx.userId)) return noAccess();
-    const summary = this.trips.getTripSummary(tripId, ctx.userId);
+    const summary = this.readModel.getTripSummary(tripId, ctx.userId);
     if (!summary) return noAccess();
     const R = canReadTrips(ctx.scopes);
     // Addon availability gates
@@ -275,7 +279,7 @@ export class TripsMcp {
     if (!this.trips.canAccessTrip(tripId, ctx.userId)) return noAccess();
     const ownerRow = this.trips.getOwner(tripId);
     if (!ownerRow) return noAccess();
-    const { owner, members } = this.trips.listMembers(tripId, ownerRow.user_id);
+    const { owner, members } = this.members.listMembers(tripId, ownerRow.user_id);
     return ok({ owner, members });
   }
 
@@ -296,7 +300,7 @@ export class TripsMcp {
     if (!ownerRow || ownerRow.user_id !== ctx.userId)
       return { content: [{ type: 'text' as const, text: 'Only the trip owner can add members.' }], isError: true };
     try {
-      const result = this.trips.addMember(tripId, identifier, ownerRow.user_id, ctx.userId);
+      const result = this.members.addMember(tripId, identifier, ownerRow.user_id, ctx.userId);
       safeBroadcast(tripId, 'member:added', { member: result.member });
       return ok({ member: result.member });
     } catch (err) {
@@ -321,7 +325,7 @@ export class TripsMcp {
     const ownerRow = this.trips.getOwner(tripId);
     if (!ownerRow || ownerRow.user_id !== ctx.userId)
       return { content: [{ type: 'text' as const, text: 'Only the trip owner can remove members.' }], isError: true };
-    this.trips.removeMember(tripId, memberId);
+    this.members.removeMember(tripId, memberId);
     safeBroadcast(tripId, 'member:removed', { userId: memberId });
     return ok({ success: true });
   }
@@ -407,7 +411,7 @@ export class TripsMcp {
     if (id === null || !this.trips.canAccessTrip(id, ctx.userId)) return accessDenied(uri.href);
     const ownerRow = this.trips.getOwner(id);
     if (!ownerRow) return accessDenied(uri.href);
-    const { owner, members } = this.trips.listMembers(id, ownerRow.user_id);
+    const { owner, members } = this.members.listMembers(id, ownerRow.user_id);
     return jsonContent(uri.href, { owner, members });
   }
 
@@ -425,7 +429,7 @@ export class TripsMcp {
     if (!this.trips.canAccessTrip(tripId, ctx.userId)) {
       return { messages: [{ role: 'user' as const, content: { type: 'text' as const, text: 'Trip not found or access denied.' } }] };
     }
-    const summary = this.trips.getTripSummary(tripId, ctx.userId);
+    const summary = this.readModel.getTripSummary(tripId, ctx.userId);
     if (!summary) {
       return { messages: [{ role: 'user' as const, content: { type: 'text' as const, text: 'Trip not found.' } }] };
     }
