+++ server/tests/unit/trips/trips.rpc.test.ts
@@ -19,6 +19,7 @@ import type { TripsService } from '../../../src/nest/trips/trips.service';
 import type { ReservationsService } from '../../../src/nest/reservations/reservations.service';
 import type { DaysService } from '../../../src/nest/days/days.service';
 import type { AccommodationsService } from '../../../src/nest/accommodations/accommodations.service';
+import type { TripMembersService } from '../../../src/nest/trip-members/trip-members.service';
 import type { TripMembershipService } from '../../../src/nest/trip-membership/trip-membership.service';
 import type { RealtimeService } from '../../../src/nest/realtime/realtime.service';
 import type { DatabaseService } from '../../../src/nest/database/database.service';
@@ -50,6 +51,10 @@ export function build(opts: { allow?: (action: string) => boolean; updateThrows?
   const accommodations = {
     list: vi.fn(() => [{ id: 11 }]),
   } as unknown as AccommodationsService & Record<string, ReturnType<typeof vi.fn>>;
+  // Membership left TripsService for its own domain.
+  const roster = {
+    removeMember: vi.fn(),
+  } as unknown as TripMembersService & Record<string, ReturnType<typeof vi.fn>>;
   const membership = { joinTripAsMember: vi.fn(() => ({ joined: true })) } as unknown as TripMembershipService & Record<string, ReturnType<typeof vi.fn>>;
   const db = {
     canAccessTrip: vi.fn((tripId: number, userId: number) => (tripId === 1 && userId === 42 ? { id: 1, user_id: 42 } : undefined)),
@@ -65,10 +70,10 @@ export function build(opts: { allow?: (action: string) => boolean; updateThrows?
   const permissions = { checkPermission: vi.fn((a: string) => (opts.allow ? opts.allow(a) : true)) } as unknown as PermissionsService;
   const realtime = { broadcast: vi.fn() } as unknown as RealtimeService & { broadcast: ReturnType<typeof vi.fn> };
   const guards = new PluginGuards(db, permissions, { isAddonEnabled: vi.fn(() => true) } as unknown as AddonsService);
-  const rpc = new TripsRpc(trips, reservations, days, membership, db, realtime, guards, accommodations);
+  const rpc = new TripsRpc(trips, reservations, days, membership, db, realtime, guards, accommodations, roster);
   const host = (...grants: string[]) =>
     new PluginRpcHost('p', new Set(grants.length ? grants : ALL_TRIP_GRANTS), makeDeps(), createTestPluginRegistry([rpc]));
-  return { trips, reservations, days, accommodations, membership, realtime, permissions, host };
+  return { trips, reservations, days, accommodations, roster, membership, realtime, permissions, host };
 }
 
 describe('TripsRpc reads', () => {
@@ -192,13 +197,13 @@ describe('TripsRpc membership', () => {
     // The stub reports user 42 as the owner of trip 1.
     const res = (await f.host().dispatch(req('trips.removeMember', { tripId: 1, userId: 42 }), 42)) as RpcError;
     expect(res.error.message).toBe('cannot remove the trip owner');
-    expect(f.trips.removeMember).not.toHaveBeenCalled();
+    expect(f.roster.removeMember).not.toHaveBeenCalled();
   });
 
   it('TRIPS-RPC-017 a non-owner member is removed', async () => {
     const f = build();
     expect((await f.host().dispatch(req('trips.removeMember', { tripId: 1, userId: 7 }), 42)).ok).toBe(true);
-    expect(f.trips.removeMember).toHaveBeenCalledWith(1, 7);
+    expect(f.roster.removeMember).toHaveBeenCalledWith(1, 7);
   });
 
   it('TRIPS-RPC-018 without member_manage nothing changes', async () => {
